Tees have been removed, the original backend is run with fd:s pointing directly to the proper files rather than having to be read and written by tea4cups, and a few small bug fixes and cleanups are also included.
Full list of changes:

* Remove the Popen4ForCUPS class
fork() and execve() directly in runOriginalBackend() instead.

* Added exception handling to logInfo()
When CUPS cancels a job before tea4cups is completely done stderr is
closed and writing to it would raise an IOError causing tea4cups to
abort without a chance to clean up.
  
* Rewrote most of runBranches() to remove tees and clean up a bit
Use retcode for consistency throughout program. Proper error
handling, falling through and doing the right thing whenever
something exits nonzero, removed the shared pipe and use stdin/stdout
instead. Make sure they aren't closed when all the pipes are closed.

* Changed stdioRedirSystem() to return actual exit code of command..
..instead of the encoded value returned by waitpid(), or -1 if the  
command didn't exit with an exit code. (E.g. killed by a signal.)  
Adjusted all places using the return value accordingly. Also fixed a
bug, if the execl() failed, the child would exit 1. Changed that to
-1 which has special meaning. (prehook cancels the job)

* Fixed indentation mistake and clean a condition in runCommand()
I use two spaces and had an incorrect indentation in the previous
patch. :) Now that tees are removed, == prehook is nicer than !=
posthook.

* Removed tee handling from runCommands() and some cleanup   
This includes moving the call to runOriginalBackend() up one level 
into runBranches(). Changed to os._exit() for forked hooks as
recommended by docs.python.org. sys.exit() is only supposed to be   
used in the parent process. Finally, the child pid for forked hooks
is already known in the waitpid loop (in pid, from the pids dict) so
I removed the variable childpid.

* Removed unregisterFileNo() and formatFileEvents()
No longer used by runOriginalBackend() and not used anywhere else.

* Rewrote runOriginalBackend() per previous no-copy theory
Fork, open dataFile and dup2() if needed, execve original backend. 
Log an error and exit -1 if execve() fails. Simplified reaping of the
original backend but it still has to keep track of SIGTERM and pass
it on if received during waitpid(). Log an error message with the
exit code of the original backend if nonzero.